refactor(onboarding): extract utils
