+++ frontend/src/scenes/onboarding/onboardingLogic.tsx
@@ -21,6 +21,7 @@ import { sidePanelStateLogic } from '~/layout/navigation-3000/sidepanel/sidePane
 import { ProductKey } from '~/queries/schema/schema-general'
 import { Breadcrumb, OnboardingProduct, OnboardingStepKey } from '~/types'
 
+import { arraysEqual, parseProductsParam, stepKeyToTitle } from './onboardingFlowUtils'
 import type { onboardingLogicType } from './onboardingLogicType'
 import { postOnboardingModalLogic } from './postOnboardingModalLogic'
 import { appendSharedTrailingSteps } from './sharedSteps'
@@ -35,54 +36,6 @@ export interface OnboardingStepComponentType<P = object> extends React.FC<P> {
 
 export interface OnboardingLogicProps {
     onCompleteOnboarding?: (key: ProductKey) => void
-}
-
-const STEP_KEY_TITLE_OVERRIDES: Partial<Record<OnboardingStepKey, string>> = {
-    [OnboardingStepKey.LINK_DATA]: 'Import data',
-}
-
-export const stepKeyToTitle = (stepKey?: OnboardingStepKey): undefined | string => {
-    if (!stepKey) {
-        return undefined
-    }
-    if (STEP_KEY_TITLE_OVERRIDES[stepKey]) {
-        return STEP_KEY_TITLE_OVERRIDES[stepKey]
-    }
-    return stepKey
-        .split('_')
-        .map((part, i) => (i == 0 ? part[0].toUpperCase() + part.substring(1) : part))
-        .join(' ')
-}
-
-const MAX_WITH_PRODUCTS = 16
-// Defense-in-depth: hard cap on the intermediate array allocated by `String.split`
-// before dedupe/filter. Browsers cap URL length already, but bounding the worst-case
-// allocation costs us nothing.
-const MAX_RAW_PRODUCT_TOKENS = 64
-
-const parseProductsParam = (raw: unknown): ProductKey[] => {
-    const value = typeof raw === 'string' ? raw : ''
-    if (!value) {
-        return []
-    }
-    // Dedupe BEFORE truncation so a malicious or buggy producer can't push valid keys
-    // out of the resulting array by stuffing the head with duplicates.
-    const unique = Array.from(new Set(value.split(',').slice(0, MAX_RAW_PRODUCT_TOKENS)))
-    return unique
-        .filter((k: string) => Object.hasOwn(availableOnboardingProducts, k))
-        .slice(0, MAX_WITH_PRODUCTS) as ProductKey[]
-}
-
-const arraysEqual = <T,>(a: readonly T[], b: readonly T[]): boolean => {
-    if (a.length !== b.length) {
-        return false
-    }
-    for (let i = 0; i < a.length; i++) {
-        if (a[i] !== b[i]) {
-            return false
-        }
-    }
-    return true
 }
 
 export const onboardingLogic = kea<onboardingLogicType>([
